Tour Description displayed, added Map, Log, Highlights-Tabs, AlertBox when deleting Tour

+++ FXApp/src/main/java/com/glatzerkratzer/tourplanner/view/TourOverviewController.java
@@ -4,7 +4,11 @@ import com.glatzerkratzer.tourplanner.model.TourItem;
 import com.glatzerkratzer.tourplanner.viewmodel.TourOverviewViewModel;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ListView;
+
+import java.util.Optional;
 
 public class TourOverviewController {
     @FXML
@@ -40,6 +44,12 @@ public class TourOverviewController {
     }
 
     public void onButtonRemove(ActionEvent actionEvent) {
-        tourOverviewViewModel.deleteTour(tourItemList.getSelectionModel().getSelectedItem());
+        Alert confirmationBox = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure you want to delete selected Tour?");
+        confirmationBox.setTitle("Delete Tour");
+
+        Optional<ButtonType> response = confirmationBox.showAndWait();
+        if (response.get() == ButtonType.OK) {
+            tourOverviewViewModel.deleteTour(tourItemList.getSelectionModel().getSelectedItem());
+        }
     }
 }
